requirements.txt
# Supabase and Database
supabase
psycopg2-binary
python-dotenv

# UI
Pillow

# Security
bcrypt
